requirements.txt
fastapi
uvicorn
websockets
opencv-python-headless
numpy
nudenet
faster-whisper
detoxify
torch
huggingface-hub
python-multipart
pydantic
aiofiles
PyJWT
